requirements.txt
semantic-kernel==1.34.0
mcp==1.9.4
fastapi==0.115.13
uvicorn==0.34.3
aiohttp==3.12.13
azure-kusto-data==5.0.4
azure-ai-ml==1.27.1
azure-ai-projects==1.0.0b11
azure-ai-resources==1.0.0b9
pyodbc==5.2.0